// src/types/components/dashboard/dashboard1.ts
/*Revenue Projects*/
type RevenueProjects = {
    img: string;
    designation: string;
    projectname: string;
    statuscolor: string;
    statustext: string;
    money: string;
    userID: string;
    name: string;
    amount: string;
    status: string;
    transactionNo: string;
    userType: string;
    gender: string;
    city: string;
    email: string;
    username: string;
    mobileNo: string;
};

/*Performance*/
type Performance = {
    icon: string;
    subtitle: string;
    title: string;
    color: string;
};

export type { RevenueProjects, Performance };